Add 'tuple_utilities' for traits/accessors
This adds a new 'tuple_utilities' file for adding a declaration for
tuple access, along with the specializable traits 'tuple_element' and
'tuple_size'
